Refactorings, keeping deleted regions in worker status record, recomputing from oplog when needed
