+++ golem-common/src/model/mod.rs
@@ -25,20 +25,19 @@ use bincode::enc::write::Writer;
 use bincode::enc::Encoder;
 use bincode::error::{DecodeError, EncodeError};
 use bincode::{BorrowDecode, Decode, Encode};
-use bytes::Bytes;
 use derive_more::FromStr;
 use poem_openapi::registry::{MetaSchema, MetaSchemaRef};
 use poem_openapi::types::{ParseFromJSON, ParseFromParameter, ParseResult, ToJSON};
 use poem_openapi::{Enum, Object};
-use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize, Serializer};
 use serde_json::Value;
 use uuid::Uuid;
 
+use crate::model::jumps::DeletedRegions;
 use crate::newtype_uuid;
-use crate::serialization::{
-    deserialize_with_version, serialize, try_deserialize, SERIALIZATION_VERSION_V1,
-};
+
+pub mod jumps;
+pub mod oplog;
 
 newtype_uuid!(
     TemplateId,
@@ -509,7 +508,6 @@ pub struct WorkerMetadata {
 }
 
 impl WorkerMetadata {
-    #[allow(unused)] // TODO
     pub fn default(worker_id: VersionedWorkerId, account_id: AccountId) -> WorkerMetadata {
         WorkerMetadata {
             worker_id,
@@ -521,9 +519,14 @@ impl WorkerMetadata {
     }
 }
 
+/// Contains status information about a worker according to a given oplog index.
+/// This status is just cached information, all fields must be computable by the oplog alone.
+/// By having an associated oplog_idx, the cached information can be used together with the
+/// tail of the oplog to determine the actual status of the worker.
 #[derive(Clone, Debug, Serialize, Deserialize, Encode, Decode)]
 pub struct WorkerStatusRecord {
     pub status: WorkerStatus,
+    pub deleted_regions: DeletedRegions,
     pub oplog_idx: u64,
 }
 
@@ -531,6 +534,7 @@ impl Default for WorkerStatusRecord {
     fn default() -> Self {
         WorkerStatusRecord {
             status: WorkerStatus::Idle,
+            deleted_regions: DeletedRegions::new(),
             oplog_idx: 0,
         }
     }
@@ -604,29 +608,6 @@ impl From<WorkerStatus> for i32 {
             WorkerStatus::Failed => 5,
             WorkerStatus::Exited => 6,
         }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::worker::WorkerMetadata> for WorkerMetadata {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::WorkerMetadata,
-    ) -> Result<Self, Self::Error> {
-        let worker_id: WorkerId = value.worker_id.ok_or("Missing worker_id")?.try_into()?;
-        Ok(Self {
-            worker_id: VersionedWorkerId {
-                worker_id,
-                template_version: value.template_version,
-            },
-            args: value.args,
-            env: value.env.into_iter().collect(),
-            account_id: value.account_id.ok_or("Missing account_id")?.into(),
-            last_known_status: WorkerStatusRecord {
-                status: value.status.try_into()?,
-                oplog_idx: 0,
-            },
-        })
     }
 }
 
@@ -732,214 +713,6 @@ impl Display for AccountId {
     }
 }
 
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Encode, Decode)]
-pub struct Jump {
-    pub target_oplog_idx: u64,
-    pub source_oplog_idx: u64,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Encode, Decode)]
-pub enum OplogEntry {
-    /// The worker invoked a host function
-    ImportedFunctionInvoked {
-        timestamp: Timestamp,
-        function_name: String,
-        response: Vec<u8>,
-        wrapped_function_type: WrappedFunctionType,
-    },
-    /// The worker has been invoked
-    ExportedFunctionInvoked {
-        timestamp: Timestamp,
-        function_name: String,
-        request: Vec<u8>,
-        invocation_key: Option<InvocationKey>,
-        calling_convention: Option<CallingConvention>,
-    },
-    /// The worker has completed an invocation
-    ExportedFunctionCompleted {
-        timestamp: Timestamp,
-        response: Vec<u8>,
-        consumed_fuel: i64,
-    },
-    /// Promise created
-    CreatePromise {
-        timestamp: Timestamp,
-        promise_id: PromiseId,
-    },
-    /// Promise completed
-    CompletePromise {
-        timestamp: Timestamp,
-        promise_id: PromiseId,
-        data: Vec<u8>,
-    },
-    /// Worker suspended
-    Suspend { timestamp: Timestamp },
-    /// Worker failed
-    Error { timestamp: Timestamp },
-    /// Debug log entry
-    Debug {
-        timestamp: Timestamp,
-        message: String,
-    },
-    /// Marker entry added when get-oplog-index is called from the worker, to make the jumping behavior
-    /// more predictable.
-    NoOp { timestamp: Timestamp },
-    /// The worker needs to recover up to the given target oplog index and continue running from
-    /// the source oplog index from there. There can be multiple active jumps during replay.
-    Jump {
-        timestamp: Timestamp,
-        jumps: Vec<Jump>,
-    },
-}
-
-impl OplogEntry {
-    pub fn imported_function_invoked<R: Encode>(
-        timestamp: Timestamp,
-        function_name: String,
-        response: &R,
-        wrapped_function_type: WrappedFunctionType,
-    ) -> Result<OplogEntry, String> {
-        let serialized_response = serialize(response)?.to_vec();
-
-        Ok(OplogEntry::ImportedFunctionInvoked {
-            timestamp,
-            function_name,
-            response: serialized_response,
-            wrapped_function_type,
-        })
-    }
-
-    pub fn exported_function_invoked<R: Encode>(
-        timestamp: Timestamp,
-        function_name: String,
-        request: &R,
-        invocation_key: Option<InvocationKey>,
-        calling_convention: Option<CallingConvention>,
-    ) -> Result<OplogEntry, String> {
-        let serialized_request = serialize(request)?.to_vec();
-        Ok(OplogEntry::ExportedFunctionInvoked {
-            timestamp,
-            function_name,
-            request: serialized_request,
-            invocation_key,
-            calling_convention,
-        })
-    }
-
-    pub fn exported_function_completed<R: Encode>(
-        timestamp: Timestamp,
-        response: &R,
-        consumed_fuel: i64,
-    ) -> Result<OplogEntry, String> {
-        let serialized_response = serialize(response)?.to_vec();
-        Ok(OplogEntry::ExportedFunctionCompleted {
-            timestamp,
-            response: serialized_response,
-            consumed_fuel,
-        })
-    }
-
-    pub fn response<T: DeserializeOwned + Decode>(&self) -> Result<Option<T>, String> {
-        match &self {
-            OplogEntry::ImportedFunctionInvoked { response, .. } => {
-                let response_bytes: Bytes = Bytes::copy_from_slice(response);
-
-                // In the v1 serialization format we did not have version prefix in the payloads.
-                // We can assume though that if the payload starts with 2, it is serialized with the
-                // v2 format because neither JSON nor protobuf (the two payload formats used in v1 for payloads)
-                // start with 2 (This was verified with a simple test ValProtobufPrefixByteValidation).
-                // So if the first byte is not 1 or 2 we assume it is a v1 format and deserialize it as JSON.
-                match try_deserialize(&response_bytes)? {
-                    Some(result) => Ok(Some(result)),
-                    None => Ok(Some(deserialize_with_version(
-                        &response_bytes,
-                        SERIALIZATION_VERSION_V1,
-                    )?)),
-                }
-            }
-            OplogEntry::ExportedFunctionCompleted { response, .. } => {
-                let response_bytes: Bytes = Bytes::copy_from_slice(response);
-
-                // See the comment above for the explanation of this logic
-                match try_deserialize(&response_bytes)? {
-                    Some(result) => Ok(Some(result)),
-                    None => Ok(Some(deserialize_with_version(
-                        &response_bytes,
-                        SERIALIZATION_VERSION_V1,
-                    )?)),
-                }
-            }
-            _ => Ok(None),
-        }
-    }
-
-    pub fn payload_as_val_array(
-        &self,
-    ) -> Result<Option<Vec<golem_wasm_rpc::protobuf::Val>>, String> {
-        // This is a special case of a possible generic request() accessor, because in v1 the only
-        // data type we serialized was Vec<Val> and it was done in a special way (every element serialized
-        // via protobuf separately, then an array of byte arrays serialized into JSON)
-        match &self {
-            OplogEntry::ExportedFunctionInvoked {
-                function_name,
-                request,
-                ..
-            } => {
-                let request_bytes: Bytes = Bytes::copy_from_slice(request);
-                self.try_decode_val_array_payload(function_name, &request_bytes)
-            }
-            OplogEntry::ExportedFunctionCompleted { response, .. } => {
-                let response_bytes: Bytes = Bytes::copy_from_slice(response);
-                self.try_decode_val_array_payload("?", &response_bytes)
-            }
-            _ => Ok(None),
-        }
-    }
-
-    fn try_decode_val_array_payload(
-        &self,
-        function_name: &str,
-        payload: &Bytes,
-    ) -> Result<Option<Vec<golem_wasm_rpc::protobuf::Val>>, String> {
-        match try_deserialize(payload)? {
-            Some(result) => Ok(Some(result)),
-            None => {
-                let deserialized_array: Vec<Vec<u8>> = serde_json::from_slice(payload)
-                    .unwrap_or_else(|err| {
-                        panic!(
-                            "Failed to deserialize oplog payload: {:?}: {err}",
-                            std::str::from_utf8(payload).unwrap_or("???")
-                        )
-                    });
-                let function_input = deserialized_array
-                    .iter()
-                    .map(|serialized_value| {
-                        <golem_wasm_rpc::protobuf::Val as prost::Message>::decode(serialized_value.as_slice())
-                            .unwrap_or_else(|err| panic!("Failed to deserialize function input {:?} for {function_name}: {err}", serialized_value))
-                    })
-                    .collect::<Vec<golem_wasm_rpc::protobuf::Val>>();
-                Ok(Some(function_input))
-            }
-        }
-    }
-
-    pub fn jump(timestamp: Timestamp, jumps: Vec<Jump>) -> OplogEntry {
-        OplogEntry::Jump { timestamp, jumps }
-    }
-
-    pub fn nop(timestamp: Timestamp) -> OplogEntry {
-        OplogEntry::NoOp { timestamp }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, Encode, Decode)]
-pub enum WrappedFunctionType {
-    ReadLocal,
-    WriteLocal,
-    ReadRemote,
-    WriteRemote,
-}
-
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct ParsedFunctionName {
     pub interface: Option<String>,
@@ -1003,12 +776,10 @@ pub fn parse_function_name(name: &str) -> ParsedFunctionName {
 
 #[cfg(test)]
 mod tests {
-    use crate::model::{parse_function_name, AccountId};
-    use crate::model::{CallingConvention, InvocationKey, Timestamp};
-    use crate::model::{OplogEntry, WrappedFunctionType};
     use bincode::{Decode, Encode};
-    use golem_wasm_rpc::protobuf::{val, Val, ValResult};
     use serde::{Deserialize, Serialize};
+
+    use crate::model::{parse_function_name, AccountId};
 
     #[test]
     fn parse_function_name_global() {
@@ -1130,170 +901,5 @@ mod tests {
         };
         let json = serde_json::to_string(&example).unwrap();
         assert_eq!(json, "{\"account_id\":\"account-1\"}");
-    }
-
-    #[test]
-    fn oplog_entry_imported_function_invoked_payload_roundtrip() {
-        let timestamp = Timestamp::now_utc();
-        let entry = OplogEntry::imported_function_invoked(
-            timestamp,
-            "function_name".to_string(),
-            &("example payload".to_string()),
-            WrappedFunctionType::ReadLocal,
-        )
-        .unwrap();
-
-        if let OplogEntry::ImportedFunctionInvoked { response, .. } = &entry {
-            assert_eq!(response.len(), 17);
-        } else {
-            unreachable!()
-        }
-
-        let response = entry.response::<String>().unwrap().unwrap();
-
-        assert_eq!(response, "example payload");
-    }
-
-    #[test]
-    fn oplog_entry_imported_function_invoked_payload_v1() {
-        let timestamp = Timestamp::now_utc();
-        let entry = OplogEntry::ImportedFunctionInvoked {
-            timestamp,
-            function_name: "function_name".to_string(),
-            response: serde_json::to_vec("example payload").unwrap(),
-            wrapped_function_type: WrappedFunctionType::ReadLocal,
-        };
-
-        let response = entry.response::<String>().unwrap().unwrap();
-
-        assert_eq!(response, "example payload");
-    }
-
-    #[test]
-    fn oplog_entry_exported_function_invoked_payload_roundtrip() {
-        let timestamp = Timestamp::now_utc();
-
-        let val1 = Val {
-            val: Some(val::Val::Result(Box::new(ValResult {
-                discriminant: 0,
-                value: Some(Box::new(Val {
-                    val: Some(val::Val::U64(10)),
-                })),
-            }))),
-        };
-        let entry = OplogEntry::exported_function_invoked(
-            timestamp,
-            "function_name".to_string(),
-            &vec![val1.clone()],
-            Some(InvocationKey {
-                value: "invocation_key".to_string(),
-            }),
-            Some(CallingConvention::Stdio),
-        )
-        .unwrap();
-
-        if let OplogEntry::ExportedFunctionInvoked { request, .. } = &entry {
-            assert_eq!(request.len(), 9);
-        } else {
-            unreachable!()
-        }
-
-        let request = entry.payload_as_val_array().unwrap().unwrap();
-
-        assert_eq!(request, vec![val1]);
-    }
-
-    #[test]
-    fn oplog_entry_exported_function_invoked_payload_v1() {
-        let timestamp = Timestamp::now_utc();
-
-        let val1 = Val {
-            val: Some(val::Val::Result(Box::new(ValResult {
-                discriminant: 0,
-                value: Some(Box::new(Val {
-                    val: Some(val::Val::U64(10)),
-                })),
-            }))),
-        };
-        let val1_bytes = prost::Message::encode_to_vec(&val1);
-        let request_bytes = serde_json::to_vec(&vec![val1_bytes]).unwrap();
-
-        let entry = OplogEntry::ExportedFunctionInvoked {
-            timestamp,
-            function_name: "function_name".to_string(),
-            request: request_bytes,
-            invocation_key: Some(InvocationKey {
-                value: "invocation_key".to_string(),
-            }),
-            calling_convention: Some(CallingConvention::Stdio),
-        };
-
-        let request = entry.payload_as_val_array().unwrap().unwrap();
-        assert_eq!(request, vec![val1]);
-    }
-
-    #[test]
-    fn oplog_entry_exported_function_completed_roundtrip() {
-        let timestamp = Timestamp::now_utc();
-
-        let val1 = Val {
-            val: Some(val::Val::Result(Box::new(ValResult {
-                discriminant: 0,
-                value: Some(Box::new(Val {
-                    val: Some(val::Val::U64(10)),
-                })),
-            }))),
-        };
-        let val2 = Val {
-            val: Some(val::Val::String("something".to_string())),
-        };
-
-        let entry = OplogEntry::exported_function_completed(
-            timestamp,
-            &vec![val1.clone(), val2.clone()],
-            1_000_000_000,
-        )
-        .unwrap();
-
-        if let OplogEntry::ExportedFunctionCompleted { response, .. } = &entry {
-            assert_eq!(response.len(), 21);
-        } else {
-            unreachable!()
-        }
-
-        let response = entry.payload_as_val_array().unwrap().unwrap();
-
-        assert_eq!(response, vec![val1, val2]);
-    }
-
-    #[test]
-    fn oplog_entry_exported_function_completed_v1() {
-        let timestamp = Timestamp::now_utc();
-
-        let val1 = Val {
-            val: Some(val::Val::Result(Box::new(ValResult {
-                discriminant: 0,
-                value: Some(Box::new(Val {
-                    val: Some(val::Val::U64(10)),
-                })),
-            }))),
-        };
-        let val1_bytes = prost::Message::encode_to_vec(&val1);
-        let val2 = Val {
-            val: Some(val::Val::String("something".to_string())),
-        };
-        let val2_bytes = prost::Message::encode_to_vec(&val2);
-
-        let response_bytes = serde_json::to_vec(&vec![val1_bytes, val2_bytes]).unwrap();
-
-        let entry = OplogEntry::ExportedFunctionCompleted {
-            timestamp,
-            response: response_bytes,
-            consumed_fuel: 1_000_000_000,
-        };
-
-        let response = entry.payload_as_val_array().unwrap().unwrap();
-
-        assert_eq!(response, vec![val1, val2]);
     }
 }
