+++ golem-worker-executor-base/src/durable_host/clocks/monotonic_clock.rs
@@ -19,7 +19,7 @@ use crate::durable_host::serialized::SerializableError;
 use crate::durable_host::{Durability, DurableWorkerCtx};
 use crate::metrics::wasm::record_host_function_call;
 use crate::workerctx::WorkerCtx;
-use golem_common::model::WrappedFunctionType;
+use golem_common::model::oplog::WrappedFunctionType;
 use wasmtime_wasi::preview2::bindings::wasi::clocks::monotonic_clock::{
     Duration, Host, Instant, Pollable,
 };
