+++ golem-worker-executor-base/src/durable_host/keyvalue/eventual_batch.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use async_trait::async_trait;
-use golem_common::model::WrappedFunctionType;
+use golem_common::model::oplog::WrappedFunctionType;
 use wasmtime::component::Resource;
 use wasmtime_wasi::preview2::{ResourceTableError, WasiView};
