+++ golem-worker-executor-base/src/durable_host/random/insecure_seed.rs
@@ -18,7 +18,7 @@ use crate::durable_host::serialized::SerializableError;
 use crate::durable_host::{Durability, DurableWorkerCtx};
 use crate::metrics::wasm::record_host_function_call;
 use crate::workerctx::WorkerCtx;
-use golem_common::model::WrappedFunctionType;
+use golem_common::model::oplog::WrappedFunctionType;
 use wasmtime_wasi::preview2::bindings::wasi::random::insecure_seed::Host;
 
 #[async_trait]
