+++ golem-worker-executor-base/src/durable_host/wasm_rpc/mod.rs
@@ -19,7 +19,8 @@ use crate::services::rpc::{RpcDemand, RpcError};
 use crate::workerctx::WorkerCtx;
 use anyhow::anyhow;
 use async_trait::async_trait;
-use golem_common::model::{TemplateId, WorkerId, WrappedFunctionType};
+use golem_common::model::oplog::WrappedFunctionType;
+use golem_common::model::{TemplateId, WorkerId};
 use golem_wasm_rpc::golem::rpc::types::Uri;
 use golem_wasm_rpc::{HostWasmRpc, WasmRpcEntry, WitValue};
 use std::str::FromStr;
